requirements.txt
tensorflow
keras
gradio
lime
scikit-learn
seaborn
numpy
Pillow
matplotlib